config/requirements/requirements.txt
# Core Dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gitpython==3.1.43
psutil==5.9.8
pyyaml==6.0.1
pydantic==2.5.2
icalendar==5.0.12
validators==0.28.1
supervision==0.19.0
docker==7.1.0
pyparsing==3.1.2
loguru==0.7.2
lxml==5.2.2
jinja2==3.1.4
werkzeug==3.0.3

# Production Server
gunicorn==21.2.0
uvicorn[standard]==0.27.0  # Updated from 0.24.0.post1

# Database - Async Support
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL driver for sync operations
aiosqlite==0.19.0  # SQLite async driver (development)
greenlet==3.0.1  # Required for SQLAlchemy async

# Async Utilities
aiofiles==23.2.1

# Background Tasks
celery==5.3.6  # Updated from 5.3.4
redis==5.0.1
# flower==2.0.1  # Commented out - unused dependency

# Real-time Communication
flask-socketio==5.3.6
python-socketio==5.11.1  # Updated from 5.11.0
# Both eventlet and gevent for SocketIO compatibility
eventlet==0.35.2  # Latest version with security fixes
gevent==23.9.1
gevent-websocket==0.10.1

# Memory Integration
httpx==0.25.2  # For Supermemory API calls

# AI/ML Libraries - CRITICAL UPDATES
openai==1.35.0  # Updated from 1.3.5 - major security fixes
anthropic==0.25.0  # Updated from 0.7.8 - important updates

# Utilities
colorama==0.4.6  # Colored terminal output
python-dateutil==2.8.2
pytz==2023.3
beautifulsoup4==4.12.3 # For HTML parsing

# Configuration & Validation
jsonschema==4.20.0  # For config validation
marshmallow==3.20.1

# Retry & Resilience
tenacity==8.4.1
apprise==1.7.6

# Observability & Monitoring
sentry-sdk[flask,celery]==1.39.1  # For error tracking and monitoring

# Security Enhancements
python-jose[cryptography]==3.3.0  # For JWT handling
PyJWT==2.8.0 # JWT implementation for python-jose
cryptography==41.0.7  # For secure operations

# Development Tools - Move to requirements-dev.txt
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
# pre-commit==3.5.0

# Optional Performance Enhancements
# ujson==5.9.0  # Faster JSON parsing
# orjson==3.9.10  # Even faster JSON parsing
# msgpack==1.0.7  # Binary serialization
watchdog==4.0.1
more-itertools==8.12.0
